wip: move components into dedicated subfolder
