+++ src/ui/pages/CredCardDetails/components/CredContentW3c.tsx
@@ -5,15 +5,15 @@ import {
   pricetagOutline,
 } from "ionicons/icons";
 import { JsonObject } from "@aries-framework/core";
-import { i18n } from "../../../i18n";
+import { i18n } from "../../../../i18n";
 import {
   CardDetailsBlock,
   CardDetailsItem,
   CardDetailsAttributes,
-} from "../../components/CardDetailsElements";
-import { formatShortDate, formatTimeToSec } from "../../utils/formatters";
-import { W3CCredentialDetails } from "../../../core/agent/services/credentialService.types";
-import { ConnectionDetails } from "../Connections/Connections.types";
+} from "../../../components/CardDetailsElements";
+import { formatShortDate, formatTimeToSec } from "../../../utils/formatters";
+import { W3CCredentialDetails } from "../../../../core/agent/services/credentialService.types";
+import { ConnectionDetails } from "../../Connections/Connections.types";
 
 interface CredContentW3cProps {
   cardData: W3CCredentialDetails;
